+++ tests/unit/Util/XmlTest.php
@@ -8,35 +8,21 @@ use PHPUnit\Framework\TestCase;
 class XmlTest extends TestCase
 {
     /**
-     * @param string    $input
-     * @param string    $expectedOutput
-     * @param bool|null $preserveEntities
+     * @param string $input
+     * @param string $expectedOutput
      *
      * @dataProvider dataProviderForTestEscape
      */
-    public function testEscape($input, $expectedOutput, $preserveEntities = null)
+    public function testEscape($input, $expectedOutput)
     {
-        if ($preserveEntities === null) {
-            $this->assertEquals($expectedOutput, Xml::escape($input));
-        } else {
-            $this->assertEquals($expectedOutput, Xml::escape($input, $preserveEntities));
-        }
+        $this->assertEquals($expectedOutput, Xml::escape($input));
     }
 
     public function dataProviderForTestEscape()
     {
-        yield ['foo', 'foo', null];
-
-        yield ['&copy;', '&amp;copy;', null];
-        yield ['&copy;', '&amp;copy;', false];
-        yield ['&copy;', '&copy;', true];
-
-        yield ['<script>', '&lt;script&gt;', null];
-        yield ['<script>', '&lt;script&gt;', false];
-        yield ['<script>', '&lt;script&gt;', true];
-
-        yield ['&#x0000;', '&amp;#x0000;', null];
-        yield ['&#x0000;', '&amp;#x0000;', false];
-        yield ['&#x0000;', '&#x0000;', true];
+        yield ['foo', 'foo'];
+        yield ['&copy;', '&amp;copy;'];
+        yield ['<script>', '&lt;script&gt;'];
+        yield ['&#x0000;', '&amp;#x0000;'];
     }
 }